CORE-19363 move batch persistence service
